+++ R/plapply.R
@@ -3,8 +3,15 @@ plapply <- function(reps, FUN, cores=0) {
   # To make the result reproducable when parallelized, seed each run 
   seeds <- sample(1:1e7, reps, replace=FALSE)
   
-  use.parallel <- cores != 1
-  have.parallel <- requireNamespace("parallel", quietly = TRUE)
+  if (.Platform$OS.type == "unix") {
+    use.parallel <- cores != 1
+    have.parallel <- requireNamespace("parallel", quietly = TRUE)
+  } else {
+    warning("parallel not supported on Windows yet...")
+    use.parallel <- FALSE
+    have.parallel <- FALSE
+  }
+  
   if (use.parallel && have.parallel) {
     # Run in parallel
     if (cores <= 0) {