+++ NEWS.md
@@ -13,6 +13,9 @@ frailtyr
   * Added parameter to `fitfrail` to compute the SE estimates
 
 ## Bug fixes
-  * Renamed `Lambda.time` to `Lambda.times`
+  * Renamed `Lambda.time` to `Lambda.times` (not backwards compatible)
   * Improved print.fitfrail
   * Remove missing observations
+  * Improved print.summary.genfrail
+  * Added sanity checks to frailty distribution params
+  * Fixed parallel support on Windows (run in serial with a warning for now)